EES-7003 - added condition to prevent "screener progress last updated" date from being updated if no actual progress has been made for a given data set since the last successful update. This will help to spot and clear out any stalled screening attempts.

+++ src/GovUk.Education.ExploreEducationStatistics.Common/Model/Screener/DataSetScreenerProgress.cs
@@ -6,7 +6,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Common.Model.Screener;
 /// </summary>
 public record DataSetScreenerProgress
 {
-    public int PercentageComplete { get; set; }
+    public double PercentageComplete { get; set; }
 
     public string Stage { get; set; } = null!;
